refactor(frontend): split sync-progress by feature
modules/shell/sync-progress held two unrelated features under one name: the premium cloud backup indicator and the history sync progress panel. Nothing was shared between them, so the cloud backup files move to modules/shell/cloud-sync and sync-progress keeps only the progress panel it is named after.

use-monitor-watchers.ts belonged to neither. It aggregates the app's watchers and its only consumer is shell/app/use-monitor-service, so it moves next to it.

Renames plus the import lines that follow them. No behaviour change.
