+++ frontend/app/src/modules/shell/app/use-monitor-watchers.ts
