+++ frontend/app/src/modules/shell/cloud-sync/AutomaticSyncSetting.vue
