+++ frontend/app/src/modules/shell/cloud-sync/SyncUploadStatusAlert.vue
