+++ frontend/app/src/modules/shell/cloud-sync/use-sync-indicator.ts
